Improve tool changing support.
This is working towards basic automatic tool changer support.  With this, is should be possible
to write a very basic macro that allows a tool changer to load specific tools based on the value
of the 'T~' gcode instruction.  If the code says 'T3 M6' to request a tool change to tool number 3,
then with this change the number 3 will be available to the macro in the 'tool' variable.

This is enough to automate tool loading, but it's not enough to make tool *unloading* work well.
There doesn't appear to be any way to remember tool slots across macro runs (yet), so we can't
automatically put the old tool back into the slot that it came out of before loading the next tool.

Changes:

- Export tool number (from T~) into the controller context so that it can be seen from inside of macros.  It can now be seen as 'tool'.
- Stop discarding the entire line whenever M6 is seen.  This ends up throwing away the T~, so that there's no way to tell which tool the gcode file was requesting.